# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Carwash Booking API
  version: 1.1.0
  description: |
    API عمومی و حساب کاربری سامانه رزرو کارواش. احراز هویت SPA با Laravel Sanctum و کوکی session انجام می‌شود.
    تاریخ‌های query میلادی با قالب YYYY-MM-DD هستند؛ پاسخ ظرفیت و رزرو فیلدهای آماده نمایش شمسی نیز دارد.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
  - url: https://api.example.ir/api/v1
    description: Production example
...
paths:
  /auth/request-otp:
    post:
      tags: [Auth]
      summary: درخواست رمز یک‌بارمصرف
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpInput'
      responses:
        '200':
          description: کد تأیید ارسال شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: تأیید رمز و ایجاد session
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpInput'
      responses:
        '200':
          description: ورود موفق
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/AuthUser'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    sanctumCookie:
      type: apiKey
      in: cookie
      name: laravel_session
      description: |
        ابتدا GET /sanctum/csrf-cookie روی origin بک‌اند فراخوانی شود و همه درخواست‌ها با credentials ارسال شوند.
...
  responses:
    ValidationError:
      description: خطای اعتبارسنجی یا قانون کسب‌وکار
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    RateLimited:
      description: تعداد درخواست بیش از حد مجاز است
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      example:
        message: اطلاعات واردشده معتبر نیست.
        errors:
          booking_slot_id:
            - ظرفیت این بازه تکمیل شده است.
    OtpPurpose:
      type: string
      enum: [login, register, verify_mobile]
    RequestOtpInput:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
          maxLength: 20
          example: '09121234567'
        purpose:
          $ref: '#/components/schemas/OtpPurpose'
      example:
        mobile: '09121234567'
        purpose: login
    VerifyOtpInput:
      type: object
      required: [mobile, code]
      properties:
        mobile:
          type: string
          maxLength: 20
        code:
          type: string
          pattern: '^\d{6}$'
        purpose:
          $ref: '#/components/schemas/OtpPurpose'
        full_name:
          type: [string, 'null']
          maxLength: 150
      example:
        mobile: '09121234567'
        code: '123456'
        purpose: login
        full_name: علی رضایی
    AuthUser:
      type: object
      required: [id, full_name, mobile]
      properties:
        id:
          type: string
        full_name:
          type: [string, 'null']
        mobile:
          type: string
...